added comments

+++ src/Components/Tweets/Tweet.js
@@ -5,6 +5,7 @@ import FavoriteIcon from '@material-ui/icons/Favorite';
 import ChatBubbleIcon from '@material-ui/icons/ChatBubble';
 import RepeatIcon from '@material-ui/icons/Repeat';
 import AuthContext from '../../Context/AuthContext';
+import { useHistory } from 'react-router-dom';
 
 const useStyles = makeStyles({
     cardS: {
@@ -35,11 +36,13 @@ tweet = {
 
 */
 
-export default function Tweet({ tweet, id, onDelete, liked = false}) {
+export default function Tweet({ tweet, id, onDelete = null, liked = false }) {
     const classes = useStyles()
     const { auth } = useContext(AuthContext)
+    const history = useHistory()
 
-    const onTweetDelete = async() => {
+    const onTweetDelete = async(e) => {
+        e.stopPropagation()
         const ret = await fetch(`http://localhost:8000/tweet`,{
             method: 'DELETE',
             headers: {'Content-Type': 'application/json'},
@@ -49,12 +52,13 @@ export default function Tweet({ tweet, id, onDelete, liked = false}) {
             })
         })
 
-        if(ret.status===200){
+        if( onDelete != null && ret.status===200){
             onDelete(id)
         }
     }
 
-    const onTweetLike = async() => {
+    const onTweetLike = async(e) => {
+        e.stopPropagation()
         const ret = await fetch(`http://localhost:8000/tweet/likes`,{
             method: 'POST',
             headers: {'Content-Type': 'application/json'},
@@ -67,8 +71,12 @@ export default function Tweet({ tweet, id, onDelete, liked = false}) {
         const like = (await ret.json()).like
     }
 
+    const onTweetClick = async() => {
+        history.push(`${tweet.username}/status/${id}`)
+    }
+
     return (
-            <Card className={classes.cardS} variant="outlined" elevation={3} square={true}>
+            <Card className={classes.cardS} variant="outlined" elevation={3} square={true} onClick={onTweetClick}>
                 <Box p={1}>
                     <CardHeader 
                         title = {tweet.name}
@@ -77,7 +85,7 @@ export default function Tweet({ tweet, id, onDelete, liked = false}) {
                         action={
                             auth && (
                             <IconButton aria-label="delete tweet">
-                            <DeleteIcon onClick={(e) => onTweetDelete()}/>
+                            <DeleteIcon onClick={(e) => onTweetDelete(e)}/>
                             </IconButton>
                             )
                         }
@@ -111,7 +119,7 @@ export default function Tweet({ tweet, id, onDelete, liked = false}) {
                                 </Typography>
                             </Grid>
                             <Grid className={classes.item}>
-                                <IconButton onClick={(e) => onTweetLike()}>
+                                <IconButton onClick={(e) => onTweetLike(e)}>
                                     <FavoriteIcon color={liked ? "secondary" : ""}/>
                                 </IconButton> 
                                 <Typography>
